services/auth-service/requirements.txt
Django>=5.0
djangorestframework>=3.15
django-cors-headers>=4.3
PyJWT>=2.8
psycopg[binary]>=3.2
python-dotenv>=1.0
pika>=1.3
gunicorn>=21.2
